feat: introducing the category value block
We need this block to be able to display the value of the categories.

I have also done a breaking change in the blocks that we are not accepting
not the category id but the category position
It is easier for the user to look at the category positions

Also adding a method for ordered category values because sometimes they can be shuffled
It is a guarding method. Always us this method.